storage: extract some types
