+++ components/tikv_misc/src/storage_types.rs
@@ -1,0 +1,34 @@
+pub use crate::storage_key::Key;
+
+// Short value max len must <= 255.
+pub const SHORT_VALUE_MAX_LEN: usize = 64;
+pub const SHORT_VALUE_PREFIX: u8 = b'v';
+
+/// Value type which is essentially raw bytes.
+pub type Value = Vec<u8>;
+
+/// Key-value pair type.
+///
+/// The value is simply raw bytes; the key is a little bit tricky, which is
+/// encoded bytes.
+pub type KvPair = (Vec<u8>, Value);
+
+#[derive(Debug, Clone)]
+pub enum Mutation {
+    Put((Key, Value)),
+    Delete(Key),
+    Lock(Key),
+    Insert((Key, Value)), // has a constraint that key should not exist.
+}
+
+#[allow(clippy::match_same_arms)]
+impl Mutation {
+    pub fn key(&self) -> &Key {
+        match *self {
+            Mutation::Put((ref key, _)) => key,
+            Mutation::Delete(ref key) => key,
+            Mutation::Lock(ref key) => key,
+            Mutation::Insert((ref key, _)) => key,
+        }
+    }
+}
